InicioViajePendiente y DetenerViajePendiente
La simulacion, la parte de iniciar los viajes, la idea estaria, hay que
acomodarla bien para mostrala por pantalla, y ver bien como llamar al
metodo desde el boton "iniciar", la parte de detener los viajes lo
invente, no creo que funcione, pero cuando tengamos todo lo demas
terminado habria que verlo bien, y probar ambas

+++ TPJava2016/src/Transporte/Transporte.java
@@ -3,13 +3,13 @@ package transporte;
 public abstract class Transporte {
 	
 	private String patente;
-	private double velocidad;
+	private int velocidad;
 	private int ocupado;
 	protected int capacidad;
 	
 	
 
-	public Transporte(String patente, double velocidad) {
+	public Transporte(String patente, int velocidad) {
 		super();
 		this.patente = patente;
 		this.velocidad = velocidad;
@@ -41,10 +41,10 @@ public abstract class Transporte {
 		this.patente = patente;
 	}
 
-	public double getVelocidad() {
+	public int getVelocidad() {
 		return velocidad;
 	}
-	public void setVelocidad(double velocidad) {
+	public void setVelocidad(int velocidad) {
 		this.velocidad = velocidad;
 	}
